% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{aggregate_to_raster}
\alias{aggregate_to_raster}
\title{Aggregate voxel data to a raster map}
\usage{
aggregate_to_raster(vox, crs = NULL, field = "value")
}
\arguments{
\item{vox}{Output of \code{\link[=preprocess_voxelize]{preprocess_voxelize()}} or \code{\link[=add_voxel_features]{add_voxel_features()}}.}

\item{crs}{EPSG code (numeric) or \link[sf:st_crs]{sf::st_crs} object.}

\item{field}{Character, which field to rasterize (default = "value").}
}
\value{
A \link[terra:SpatRaster-class]{terra::SpatRaster}.
}
\description{
Converts voxelized 3D data (X, Y, value) into a 2D raster layer.
}
