% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\name{get_batch}
\alias{get_batch}
\title{Fetch the next batch from a loader list}
\usage{
get_batch(loader, i = 1L)
}
\arguments{
\item{loader}{Object created by \code{\link[=create_dataloader]{create_dataloader()}}.}

\item{i}{Batch index (defaults to 1).}
}
\value{
A batch (list of tensors).
}
\description{
Fetch the next batch from a loader list
}
