% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pointnet.R
\name{pointnet_model}
\alias{pointnet_model}
\title{Define a simple PointNet-style neural network}
\usage{
pointnet_model(input_dim = 3, output_dim = 2)
}
\arguments{
\item{input_dim}{Number of input features (e.g. 3 for X,Y,Z or 4 for X,Y,Z,value)}

\item{output_dim}{Number of target classes (for regression use 1)}
}
\value{
A torch nn_module
}
\description{
Define a simple PointNet-style neural network
}
