% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_read_write.R
\name{read_pointcloud}
\alias{read_pointcloud}
\title{Read or write 3D point-cloud files}
\usage{
read_pointcloud(path, crs = NULL)
}
\arguments{
\item{path}{Path to the LAS/LAZ/PLY file.}

\item{crs}{Optional EPSG code (integer or string) to assign if missing.}
}
\value{
A \link[lidR:LAS-class]{lidR::LAS} object with header information.
}
\description{
Wrapper around \pkg{lidR} and \pkg{rlas} readers supporting LAS/LAZ/PLY formats.
}
\examples{
pc <- read_pointcloud(system.file("extdata", "forest_plot.laz", package = "ecoPointXAI"))
}
