% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_read_write.R
\name{write_pointcloud}
\alias{write_pointcloud}
\title{Write processed LAS/LAZ file}
\usage{
write_pointcloud(pc, path)
}
\arguments{
\item{pc}{A \link[lidR:LAS-class]{lidR::LAS} object to write.}

\item{path}{Output path ending with \code{.las} or \code{.laz}.}
}
\value{
Invisibly returns TRUE if successful.
}
\description{
Wrapper around \pkg{lidR::writeLAS} ensuring valid extension and silent return.
}
\examples{
\dontrun{
pc <- read_pointcloud(system.file("extdata", "forest_plot.laz", package = "ecoPointXAI"))
tmp <- tempfile(fileext = ".laz")
write_pointcloud(pc, tmp)
}
}
